Clean compression/__init__.py
